Issue #68 refactor: scheduleId 타입 변경
일정 추가 시 rowId를 반환하기 위해 id 타입을 Long으로 변경하고 insertSchedule의
리턴 타입을 Long으로 변경.

+++ app/src/main/java/com/drunkenboys/calendarun/ui/saveschedule/SaveScheduleViewModel.kt
@@ -24,7 +24,7 @@ import com.drunkenboys.calendarun.data.calendar.entity.Calendar as CalendarEntit
 @HiltViewModel
 class SaveScheduleViewModel @Inject constructor(
     @CalendarId private val calendarId: Int,
-    @ScheduleId private val scheduleId: Int,
+    @ScheduleId private val scheduleId: Long,
     private val calendarDataSource: CalendarLocalDataSource,
     private val scheduleDataSource: ScheduleLocalDataSource
 ) : ViewModel() {
